requirements.txt
tqdm
semgrep
requests
python-dateutil